backend/requirements.txt
google-cloud-discoveryengine
google-cloud-storage
fastapi
coloredlogs
python-multipart